Refactor client code
Move things into directories to tidy stuff up
