+++ src/styles/buddypress.scss
