+++ src/styles/editor.scss
